fix(接口测试): 用例详情的响应结果展示状态码、响应时间、响应大小
--bug=1038129 --user=吕梦园【接口测试】定义-用例列表-操作点击执行-响应结果未展示状态码、响应时间、响应大小
https://www.tapd.cn/55049933/bugtrace/bugs/view/1155049933001038129
